Move experiment_path to environment var
Depending on whether this is being run locally or on a docker machine, it may be necessary to change the location into which experiments are saved. Move the experiment path to an environment var to allow it to be easily changed across files.

TODO: Move other path variables (e.g. output) to environment vars.

Also reduce the size of test files to reduce test runtime.